Feat: 스터디상세,강의상세,리스트 api 유저 북마크 여부 추가 및 스터디 지원 취소 api 구현  (#118)
* Feat: 스터디 지원 취소 구현

* Feat: 스터디 상세, 강의상세, 스터디리스트, 강의리스트 api 유저북마크 여부 추가

* Feat: 홈, 검색결과 isBookmarked 연결

+++ src/app/api/study/[id]/route.ts
@@ -1,4 +1,5 @@
 import connectDB from "@/lib/database/db";
+import { StudyBookmark } from "@/schema/bookmarkSchema";
 import { Lecture } from "@/schema/lectureSchema";
 import { Study } from "@/schema/studySchema";
 import { TeamMembers } from "@/schema/teamMemberSchema";
@@ -9,8 +10,10 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
   if (!studyId) {
     return Response.json({ error: "id 가 필요합니다." }, { status: 400 });
   }
-  connectDB();
+  await connectDB();
 
+  const token = req.headers.get("Authorization");
+  const userId = token?.split(" ")[1];
   try {
     const study = await Study.findById(studyId);
     if (!study) {
@@ -23,7 +26,8 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
     if (!isOwnerExisted) {
       return Response.json({ message: "해당 스터디의 작성자가 존재하지 않습니다." }, { status: 410 });
     }
-
+    const isBookmarked = userId ? Boolean(await StudyBookmark.findOne({ studyId, userId })) : false;
+    console.log(isBookmarked);
     const lectureId = study.lectureId;
     const lectureData = await Lecture.findById(lectureId);
     const { _id, category, online, platform, rating, title, instructor, link } = lectureData;
@@ -50,7 +54,10 @@ export async function GET(req: Request, { params }: { params: { id: string } }) 
       );
     }
 
-    return Response.json({ study, lecture: lectureResult, teamMemberList: acceptedTeamMembers }, { status: 200 });
+    return Response.json(
+      { study, lecture: lectureResult, teamMemberList: acceptedTeamMembers, isBookmarked },
+      { status: 200 },
+    );
   } catch (error: any) {
     if (error.name === "CastError") {
       return Response.json({ message: "스터디가 존재하지 않습니다." }, { status: 404 });
@@ -65,11 +72,21 @@ export async function DELETE(request: Request, { params }: { params: { id: strin
   if (!studyId) {
     return Response.json({ error: "id 가 필요합니다." }, { status: 400 });
   }
-  connectDB();
+
+  const token = request.headers.get("Authorization");
+  const userId = token?.split(" ")[1];
+  if (!userId) {
+    return Response.json({ error: "user id 가 필요합니다." }, { status: 400 });
+  }
+  await connectDB();
   try {
     const study = await Study.findById(studyId);
     if (!study) {
       return Response.json({ message: "스터디가 존재하지 않습니다." }, { status: 404 });
+    }
+
+    if (study.ownerId !== userId) {
+      return Response.json({ message: "권한이 없습니다." }, { status: 403 });
     }
     await Study.findByIdAndDelete(studyId);
